Refactored ProjectsSection structure - Moved card appear animation inside ProjectCard; - Extracted blocks out of ProjectsSection.tsx into more manageable units; - Refactored ProjectCard; - Use relative imports inside ProjectCard;
