+++ src/components/sections/ProjectsSection/animations/animateCardsList.ts
@@ -1,13 +1,10 @@
 import Project from "@/lib/models/Project";
-import { ProjectsSectionReferences } from "@/components/sections/ProjectsSection/types";
 import animateCardStacking from "@/components/sections/ProjectsSection/animations/animateCardStacking";
 import animateRootGradient from "@/components/sections/ProjectsSection/animations/animateRootGradient";
-import animateCardContentAppear from "@/components/sections/ProjectsSection/animations/animateCardContentAppear";
 
-export default function animateProjectCards(
+export default function animateCardsList(
     cards: NodeListOf<Element>,
     projects: Project[],
-    references: ProjectsSectionReferences,
     scope: HTMLElement
 ) {
     const lastCard = cards[cards.length - 1];
@@ -19,11 +16,5 @@ export default function animateProjectCards(
 
         animateCardStacking(card, index, indexFromEnd, lastCard);
         animateRootGradient(project, prevProject, scope, card, index);
-
-        const cardContent = card.querySelector(`.${references.projectCard.content.className}`);
-        const techLabels = card.querySelectorAll(`.${references.projectCard.content.techLabelClassName}`);
-        if (cardContent) {
-            animateCardContentAppear(cardContent, techLabels, index);
-        }
     });
 }
