+++ src/components/sections/ProjectsSection/types.ts
@@ -1,7 +1,7 @@
 import {ProjectCardReference} from "@/components/sections/ProjectsSection/components/ProjectCard/types";
 
 export interface ProjectsSectionReferences {
-    projectsSectionId: string;
     projectsHeaderId: string;
+    projectsListId: string;
     projectCard: ProjectCardReference;
 }
